requirements.txt
aiohttp==3.9.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.8.1
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
Django==4.2.13
django-environ==0.11.2
django-phonenumber-field==7.3.0
djangorestframework==3.15.1
drfpasswordless==1.5.9
frozenlist==1.4.1
idna==3.7
multidict==6.0.5
phonenumbers==8.13.37
PyJWT==2.8.0
requests==2.31.0
sqlparse==0.5.0
twilio==9.0.5
urllib3==2.2.1
yarl==1.9.4
